requirements.txt
mysql-connector-python>=8.0.0
faker>=19.0.0
pandas>=2.0.0
matplotlib>=3.7.0
fpdf2>=2.7.0
pyyaml>=6.0
